Move decorations utils out of sourcecode api.
