+++ spyder/widgets/sourcecode/utils/decoration.py
